requirements.txt
-e git+http://github.com/Rapptz/discord.py#egg=discord.py
Flask==2.1.1
pymongo==4.1.1
python-dotenv==0.20.0
requests==2.27.1
selenium==4.1.5
webdriver_manager==3.5.4
yt_dlp==2022.1.21
dnspython==2.2.1
pystreamable==1.0.1